Move ImplementAvgBigint logic from connectors into base jdbc module
