+++ plugin/trino-mysql/src/main/java/io/trino/plugin/mysql/ImplementAvgBigint.java
@@ -13,55 +13,14 @@
  */
 package io.trino.plugin.mysql;
 
-import io.trino.matching.Capture;
-import io.trino.matching.Captures;
-import io.trino.matching.Pattern;
-import io.trino.plugin.base.expression.AggregateFunctionRule;
-import io.trino.plugin.jdbc.JdbcColumnHandle;
-import io.trino.plugin.jdbc.JdbcExpression;
-import io.trino.plugin.jdbc.JdbcTypeHandle;
-import io.trino.spi.connector.AggregateFunction;
-import io.trino.spi.expression.Variable;
-
-import java.sql.Types;
-import java.util.Optional;
-
-import static com.google.common.base.Verify.verify;
-import static io.trino.matching.Capture.newCapture;
-import static io.trino.plugin.base.expression.AggregateFunctionPatterns.basicAggregation;
-import static io.trino.plugin.base.expression.AggregateFunctionPatterns.expressionType;
-import static io.trino.plugin.base.expression.AggregateFunctionPatterns.functionName;
-import static io.trino.plugin.base.expression.AggregateFunctionPatterns.singleInput;
-import static io.trino.plugin.base.expression.AggregateFunctionPatterns.variable;
-import static io.trino.spi.type.BigintType.BIGINT;
-import static io.trino.spi.type.DoubleType.DOUBLE;
-import static java.lang.String.format;
+import io.trino.plugin.jdbc.expression.BaseImplementAvgBigint;
 
 public class ImplementAvgBigint
-        implements AggregateFunctionRule<JdbcExpression>
+        extends BaseImplementAvgBigint
 {
-    private static final Capture<Variable> INPUT = newCapture();
-
     @Override
-    public Pattern<AggregateFunction> getPattern()
+    protected String getRewriteFormatExpression()
     {
-        return basicAggregation()
-                .with(functionName().equalTo("avg"))
-                .with(singleInput().matching(
-                        variable()
-                                .with(expressionType().matching(type -> type == BIGINT))
-                                .capturedAs(INPUT)));
-    }
-
-    @Override
-    public Optional<JdbcExpression> rewrite(AggregateFunction aggregateFunction, Captures captures, RewriteContext context)
-    {
-        Variable input = captures.get(INPUT);
-        JdbcColumnHandle columnHandle = (JdbcColumnHandle) context.getAssignment(input.getName());
-        verify(aggregateFunction.getOutputType() == DOUBLE);
-
-        return Optional.of(new JdbcExpression(
-                format("avg((%s * 1.0))", context.getIdentifierQuote().apply(columnHandle.getColumnName())),
-                new JdbcTypeHandle(Types.DOUBLE, Optional.of("double"), Optional.empty(), Optional.empty(), Optional.empty(), Optional.empty())));
+        return "avg((%s * 1.0))";
     }
 }
